edit index mission report view

+++ views/mission_report/index.php
@@ -73,38 +73,46 @@
         </div>
         
         <!-- Mission Report Section -->
-        <div class="lg:w-2/3 p-6 <?php if(!$mission_report_list) echo "hidden"; ?>">
-            <h2 class="text-2xl font-bold mb-2">Mission Report</h2>
-            <div class="mb-4">
-                <p>Report summary for <?php echo $mission->name;?></p>
-                <ul class="mt-2 space-y-2">
-                    <li><span class="font-semibold">Injured:</span> 19 people</li>
-                    <li><span class="font-semibold">Died:</span> 15 people</li>
-                </ul>
-            </div>
-            <div class="grid grid-cols-1 md:grid-cols-2 gap-4">
-                <?php foreach($mission_report_list as $mission_report) : ?>
-                    <div class="bg-gray-50 p-4 rounded shadow">
-                        <div class="flex flex-row space-x-2">
-                            <h3 class="flex-auto font-semibold mb-2">Report ID[<?php echo $mission_report->id; ?>]</h3>
-                            <a 
-                                href="?controller=missionReport&action=editForm&id=<?php echo $mission_report->id; ?>" 
-                                class="inline-flex items-center text-sm font-semibold rounded-lg border border-transparent text-blue-600 hover:text-blue-800 focus:outline-none focus:text-blue-800 transition duration-200"
-                            >
-                                <i class="fa-regular fa-pen-to-square"></i>
-                            </a>
-                            <a 
-                                href="?controller=missionReport&action=deleteForm&id=<?php echo $mission_report->id; ?>" 
-                                class="inline-flex items-center text-sm font-semibold rounded-lg border border-transparent text-red-600 hover:text-red-800 focus:outline-none focus:text-red-800 transition duration-200"
-                            >
-                                <i class="fa-solid fa-trash"></i>
-                            </a>
-                        </div>
-                        <p>Created date : <?php echo date('d M Y H:i:s', strtotime($mission_report->date)); ?></p>
-                        <p class="text-wrap break-words">Detail : <?php echo $mission_report->detail; ?></p>
+         <?php if(isset($mission_report_list)) : ?>
+            <?php if($mission_report_list) : ?>
+                <div class="lg:w-2/3 p-6">
+                    <h2 class="text-2xl font-bold mb-2">Mission Report</h2>
+                    <div class="mb-4">
+                        <p>Report summary for <?php echo $mission->name;?></p>
+                        <ul class="mt-2 space-y-2">
+                            <li><span class="font-semibold">Injured:</span> 19 people</li>
+                            <li><span class="font-semibold">Died:</span> 15 people</li>
+                        </ul>
                     </div>
-                <?php endforeach; ?>
-            </div>
-        </div>
+                    <div class="grid grid-cols-1 md:grid-cols-2 gap-4">
+                        <?php foreach($mission_report_list as $mission_report) : ?>
+                            <div class="bg-gray-50 p-4 rounded shadow">
+                                <div class="flex flex-row space-x-2">
+                                    <h3 class="flex-auto font-semibold mb-2">Report ID[<?php echo $mission_report->id; ?>]</h3>
+                                    <a 
+                                        href="?controller=missionReport&action=editForm&id=<?php echo $mission_report->id; ?>" 
+                                        class="inline-flex items-center text-sm font-semibold rounded-lg border border-transparent text-blue-600 hover:text-blue-800 focus:outline-none focus:text-blue-800 transition duration-200"
+                                    >
+                                        <i class="fa-regular fa-pen-to-square"></i>
+                                    </a>
+                                    <a 
+                                        href="?controller=missionReport&action=deleteForm&id=<?php echo $mission_report->id; ?>" 
+                                        class="inline-flex items-center text-sm font-semibold rounded-lg border border-transparent text-red-600 hover:text-red-800 focus:outline-none focus:text-red-800 transition duration-200"
+                                    >
+                                        <i class="fa-solid fa-trash"></i>
+                                    </a>
+                                </div>
+                                <p>Created date : <?php echo date('d M Y H:i:s', strtotime($mission_report->date)); ?></p>
+                                <p class="text-wrap break-words">Detail : <?php echo $mission_report->detail; ?></p>
+                            </div>
+                        <?php endforeach; ?>
+                    </div>
+                </div>
+            <?php else : ?>
+                <div class="lg:w-2/3 p-6 flex justify-center items-center">
+                    <h2 class="text-2xl font-semibold">There is no mission report</h2>
+                </div>
+            <?php endif; ?>
+        <?php endif; ?>
     </div>
 </div>
